requirements.txt
# 🔐 Crypto + Signing
cryptography>=42.0.5
pyjwt[crypto]>=2.8.0

# 🧪 Validation
pydantic>=2.9.0

# 🧬 Environment management
python-dotenv>=1.0.1

# ⚡ HTTP client for agents (e.g. Apple ID, APIs)
httpx>=0.27.0

# 🛠️ CLI + scripting
# argparse is built-in for Python 3.2+

# 🧪 Testing
pytest>=8.3.0

# 📦 Optional OAuth / Apple ID agents
requests>=2.32.3
